+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 - [BREAKING]: `cells(mesh)` returns no longer tuples of integers that are indices into the mesh vertex buffer, but typed objects that should be interpreted as *local graphs*, and that can be querried for adjacency and orientation information. This is done to avoid ambiguity when considering higher order meshes and to allow core algorithms such as `connectivity` and `skeleton` to be written generically for both lowest order and higher order meshes.
+- [BREAKING]: PlotlyJS support is now an implemented as extension and patch, wireframe, and normals are retired in favour of PlotlyJS standard names mesh3d, scatter3d, and cone.
 
 ## Version 0.10